+++ core/state/state_frozen_fund.go
@@ -141,9 +141,9 @@ func (c *stateFrozenFund) punishFund(candidateAddress [20]byte) {
 			slashed := big.NewInt(0).Set(item.Value)
 			slashed.Sub(slashed, newValue)
 
-			edb.SaveEvent(int64(c.blockHeight), eventsdb.SlashEvent{
+			edb.AddEvent(int64(c.blockHeight), eventsdb.SlashEvent{
 				Address:         item.Address,
-				Amount:          slashed.String(),
+				Amount:          slashed.Bytes(),
 				Coin:            item.Coin,
 				ValidatorPubKey: item.CandidateKey,
 			})