+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
 ## 0.3.4
-*Sept 12th, 2018*
+*Sept 13th, 2018*
 
 IMPROVEMENT
 
-- [api] Optimize events
+- [api] Optimize events. WARNING! If you are using events you should re-sync blockchain from scratch.
 - [api] Refactor api
 
 ## 0.3.3